# api/openapi.yaml
openapi: 3.1.0
info:
  title: Runtrue API
  version: 0.1.0
  description: >-
    Preliminary control-plane contract. Mutation responses may include
    Idempotency-Replayed. Secret values and enrollment bearer tokens are never
    returned by read endpoints. Enrolled-runner secret and OIDC broker RPCs
    use the separately versioned `runtrue.runner.v1.RunnerControl` protobuf
    contract; they require mTLS plus a live certificate-owned Open session and
    are intentionally not modeled as HTTP endpoints here.
servers:
  - url: https://ci.example.com
security:
  - bearerAuth: []
paths:
...
  /auth/callback:
    get:
      operationId: finishGithubOauthLogin
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 8192}
        - name: state
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 1024}
      responses:
        '303': {description: Verified GitHub identity session created}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
  /auth/oidc/callback:
    get:
      operationId: finishHumanOidcLogin
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 8192}
        - name: state
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 1024}
      responses:
        '303': {description: Verified pre-provisioned identity session created}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    browserAccess:
      type: apiKey
      in: cookie
      name: runtrue_access
    browserRefresh:
      type: apiKey
      in: cookie
      name: runtrue_refresh
    browserCsrf:
      type: apiKey
      in: cookie
      name: runtrue_csrf
...
  responses:
...
    Unauthorized:
      description: Bearer or webhook authentication failed
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    ServiceUnavailable:
      description: Durable service or configured integration unavailable
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, request_id]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        request_id: {type: string}